# WrapUnsafe in hash transformations