--- a/transformations/sha1.go
+++ b/transformations/sha1.go
@@ -6,13 +6,18 @@ package transformations
 import (
 	"crypto/sha1"
 	"io"
+
+	"github.com/corazawaf/coraza/v3/internal/strings"
 )
 
 func sha1T(data string) (string, error) {
+	if len(data) == 0 {
+		return "", nil
+	}
 	h := sha1.New()
 	_, err := io.WriteString(h, data)
 	if err != nil {
 		return data, err
 	}
-	return string(h.Sum(nil)), nil
+	return strings.WrapUnsafe(h.Sum(nil)), nil
 }