After
```
BenchmarkSHA1
BenchmarkSHA1/#00
BenchmarkSHA1/#00-10         	555318814	         2.120 ns/op
BenchmarkSHA1/1234567890
BenchmarkSHA1/1234567890-10  	11032917	       108.5 ns/op
BenchmarkMD5
BenchmarkMD5/#00
BenchmarkMD5/#00-10         	563206208	         2.136 ns/op
BenchmarkMD5/1234567890
BenchmarkMD5/1234567890-10  	 7238055	       165.6 ns/op
```

Before
```
BenchmarkSHA1
BenchmarkSHA1/#00
BenchmarkSHA1/#00-10         	10593223	       110.5 ns/op
BenchmarkSHA1/1234567890
BenchmarkSHA1/1234567890-10  	10065594	       123.5 ns/op
BenchmarkMD5
BenchmarkMD5/#00
BenchmarkMD5/#00-10         	 7257741	       164.5 ns/op
BenchmarkMD5/1234567890
BenchmarkMD5/1234567890-10  	 6698254	       175.3 ns/op
```